Moved all tests into one file
